requirement.txt
absl-py==1.4.0
accelerate==0.15.0
aiofiles==22.1.0
aiohttp==3.7.4.post0
altair==4.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
asgiref==3.6.0
async-timeout==3.0.1
attrs==22.2.0
backoff==2.2.1
bentoml==1.0.13
bitsandbytes==0.35.0
black==22.12.0
build==0.10.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
circus==0.18.0
click==8.1.3
click-option-group==0.5.5
cloudpickle==2.2.1
cmake==3.25.0
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
datasets==2.8.0
deepmerge==1.1.0
Deprecated==1.2.13
diffuser==0.0.1
diffusers @ git+https://github.com/huggingface/diffusers.git@a66f2baeb782e091dde4e1e6394e46f169e5ba58
dill==0.3.6
docker-pycreds==0.4.0
einops==0.4.2
entrypoints==0.4
exceptiongroup==1.1.0
fastapi==0.89.1
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
formers==0.0.16rc391
fs==2.4.16
fsspec==2023.1.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.58.0
gradio==3.16.2
grpcio==1.51.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==1.0.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
modelcards==0.1.6
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==0.4.3
ninja==1.11.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opentelemetry-api==1.14.0
opentelemetry-exporter-otlp-proto-http==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
orjson==3.8.5
packaging==21.3
pandas==1.5.3
pathspec==0.11.0
pathtools==0.1.2
Pillow==9.4.0
pip-requirements-parser==32.0.1
pip-tools==6.12.1
platformdirs==2.6.2
prometheus-client==0.16.0
protobuf==3.20.3
psutil==5.9.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.16.0
pydantic==1.10.4
pydub==0.25.1
Pygments==2.14.0
pynvml==11.4.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyre-extensions==0.0.23
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rich==13.3.0
rsa==4.9
schema==0.7.5
sentry-sdk==1.14.0
setproctitle==1.3.2
simple-di==0.1.5
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.22.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
transformers==4.25.1
triton==2.0.0.dev20221202
typing-inspect==0.8.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
wandb==0.13.9
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
xformers @ git+https://github.com/facebookresearch/xformers.git@bc08bbc631348913a3c37b4e09832973ff93a398
xxhash==3.2.0
yarl==1.8.2
zipp==3.11.0
